requirements.txt
thordata-sdk>=0.3.1
openai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0